Test for more permissions-friendly I/O

+++ MonoGame.Framework/Desktop/Audio/OpenALSoundController.cs
@@ -50,10 +50,22 @@ namespace Microsoft.Xna.Framework.Audio
 
         private OpenALSoundController()
         {
-            // choose the right DLL!
-            File.Delete("openal32.dll");
-            if (IntPtr.Size == 8)   File.Copy("soft_oal_64.dll", "openal32.dll");
-            else                    File.Copy("soft_oal_32.dll", "openal32.dll");
+            try
+            {
+                File.SetAttributes("openal32.dll", FileAttributes.Normal);
+                File.Delete("openal32.dll");
+            }
+            catch (UnauthorizedAccessException)
+            {
+                // no big deal, we either already have the right openal32.dll file, or we're about to create it
+            }
+
+            if (!File.Exists("openal32.dll"))
+            {
+                // choose the right DLL!
+                if (IntPtr.Size == 8) File.Copy("soft_oal_64.dll", "openal32.dll");
+                else File.Copy("soft_oal_32.dll", "openal32.dll");
+            }
 
             try
             {
@@ -61,8 +73,10 @@ namespace Microsoft.Xna.Framework.Audio
             }
             catch (TypeInitializationException)
             {
-                // Badly advertised, but we're using the wrong DLL... try the other one.
+                // Badly advertised, but we're using the wrong DLL... try the other one. (and hope that it works!)
+                File.SetAttributes("openal32.dll", FileAttributes.Normal);
                 File.Delete("openal32.dll");
+
                 if (IntPtr.Size != 8)   File.Copy("soft_oal_64.dll", "openal32.dll");
                 else                    File.Copy("soft_oal_32.dll", "openal32.dll");
